feat(dashboard): split settings nav into dedicated pages
Breaks the monolithic Integrations page into focused sections:
- Navigation: nav logo, title, links, social links, footer nav
- Features: search, comments (+ Giscus config), edit link
- Analytics: Google Analytics, Umami
- GitHub: repository connection, auto-sync

Moves Syntax Mode from General to Content. Old /integrations now
redirects to /github. All new pages read from getDbConfig (raw DB
configJson) rather than the merged getConfig result.
